Issue #1163 rework ehcache distribution jars
Transactions is a standalone module, clustered jar only includes clustered stuff, depends on ehcache, is included in kit
Kit adjustments - stripped top level subdir, ehcache and 3rd party jars separated
