+++ build.gradle
@@ -59,9 +59,9 @@ subprojects {
 
   dependencies {
     if (JavaVersion.current().compareTo(JavaVersion.VERSION_1_7) >= 0) {
-      compile 'com.google.code.findbugs:annotations:3.0.0'
+      compileOnly 'com.google.code.findbugs:annotations:3.0.0'
     } else {
-      compile 'com.google.code.findbugs:annotations:2.0.3'
+      compileOnly 'com.google.code.findbugs:annotations:2.0.3'
     }
     testCompile 'junit:junit:4.11', 'org.hamcrest:hamcrest-library:1.3', 'org.mockito:mockito-core:1.9.5'
     testRuntime "org.slf4j:slf4j-simple:$parent.slf4jVersion"
