+++ clustered/client/build.gradle
@@ -17,12 +17,14 @@
 apply plugin: EhDeploy
 
 dependencies {
-  compile project(':api')
-  compile project(':xml')
+  compileOnly project(':api')
+  compileOnly project(':xml')
   compile project(':clustered:common')
   compile "org.terracotta.entities:clustered-map-client:$parent.clusteredMapVersion"
   compile "org.terracotta:entity-client-api:$parent.entityApiVersion"
 
+  testCompile project(':api')
+  testCompile project(':xml')
   testCompile(project(':clustered:server')) {
     exclude group: 'org.terracotta.internal', module: 'tc-config-parser'
   }
