+++ transactions/build.gradle
@@ -14,11 +14,21 @@
  * limitations under the License.
  */
 
+group = 'org.ehcache'
+
+apply plugin: EhOsgi
 apply plugin: EhDeploy
+apply plugin: EhPomMangle
 
 dependencies {
   compile project(':impl'), project(':xml')
   compile group: 'javax.transaction', name: 'jta', version: '1.1'
   testCompile project(path: ':core-spi-test')
   compile group: 'org.codehaus.btm', name: 'btm', version: '2.1.4'
+}
+
+// For EhPomMangle
+dependencies {
+  pomOnlyCompile "org.ehcache:ehcache:$parent.baseVersion"
+  pomOnlyProvided 'javax.transaction:jta:1.1', 'org.codehaus.btm:btm:2.1.4'
 }
